refactor: url utils get path parts, storage dto is dir check
